% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{access_token}
\alias{access_token}
\title{OAuth Access Token}
\usage{
access_token(consumer, request_token, verifier)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{request_token}{A request token, created by \code{\link{request_token}}.}

\item{verifier}{A verification code, received after authorization.}
}
\value{
An access token, an object of class \code{"token"}.
}
\description{
Get an access token.
}
\examples{
\dontrun{
# consumer key and secret below are for illustrative purposes only
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf"
)
my_request_token <- request_token(my_consumer)
my_authorization_url <- authorization_url(my_consumer, my_request_token)
browseURL(my_authorization_url)
# verification code below is for illustrative purposes only
my_access_token <- access_token(my_consumer, my_request_token, "jHuNpFdZOB8JXNzR")
}
}
