% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{authorization_url}
\alias{authorization_url}
\title{OAuth Authorization URL}
\usage{
authorization_url(consumer, request_token)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{request_token}{A request token, created by \code{\link{request_token}}.}
}
\value{
An authorization URL.
}
\description{
Get an authorization URL.
}
\examples{
\dontrun{
# consumer key and secret below are for illustrative purposes only
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf"
)
my_request_token <- request_token(my_consumer)
my_authorization_url <- authorization_url(my_consumer, my_request_token)
browseURL(my_authorization_url)
}
}
