% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{chppxml}
\alias{chppxml}
\title{CHPP XML Files}
\usage{
chppxml(query, consumer, access_token)
}
\arguments{
\item{query}{A query string, e.g. \code{"file=teamdetails&version=3.4"}.}

\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{access_token}{An access token, created by \code{\link{access_token}}.}
}
\value{
A CHPP XML file.
}
\description{
Get CHPP XML files.
}
\examples{
\dontrun{
# consumer key and secret below are for illustrative purposes only
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf"
)
my_request_token <- request_token(my_consumer)
my_authorization_url <- authorization_url(my_consumer, my_request_token)
browseURL(my_authorization_url)
# verification code below is for illustrative purposes only
my_access_token <- access_token(my_consumer, my_request_token, "jHuNpFdZOB8JXNzR")
my_chppxml <- chppxml("file=teamdetails&version=3.4", my_consumer, my_access_token)
}
}
