% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-constructors.R
\name{consumer}
\alias{consumer}
\title{OAuth Consumer}
\usage{
consumer(key, secret, challenge = FALSE, order = FALSE,
  youth = FALSE, training = FALSE, bid = FALSE)
}
\arguments{
\item{key}{Consumer key.}

\item{secret}{Consumer secret.}

\item{challenge}{Manage challenges?}

\item{order}{Set match orders?}

\item{youth}{Manage youth players?}

\item{training}{Set training?}

\item{bid}{Place bids?}
}
\value{
An object of class \code{"consumer"}.
}
\description{
Create a consumer.
}
\examples{
# consumer key and secret below are for illustrative purposes only
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf"
)
# ask for permission to manage challenges
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf",
  challenge = TRUE
)
}
