% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ht_languages}
\alias{ht_languages}
\title{Hattrick Languages}
\format{A data frame with 57 rows and 2 columns:
\describe{
  \item{\code{language_id}}{Language ID.}
  \item{\code{language_name}}{Language name.}
}}
\usage{
ht_languages
}
\description{
A data frame with available languages.
}
\keyword{datasets}
