% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{languages}
\alias{languages}
\title{Languages}
\usage{
languages(consumer, access_token)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{access_token}{An access token, created by \code{\link{access_token}}.}
}
\value{
A data frame with two variables:
\item{\code{language_id}}{Language ID.}
\item{\code{language_name}}{Language name.}
}
\description{
Download a data frame with available languages.
}
\examples{
\dontrun{
df <- languages(consumer, access_token)
}
}
