% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-parameters.R
\name{oauth_url}
\alias{oauth_url}
\title{OAuth URLs}
\usage{
oauth_url()
}
\value{
A list with seven elements:
\item{\code{request_token}}{Request token URL.}
\item{\code{authorize}}{Authorize URL.}
\item{\code{authenticate}}{Authenticate URL.}
\item{\code{access_token}}{Access token URL.}
\item{\code{check_token}}{Check token URL.}
\item{\code{invalidate_token}}{Invalidate token URL.}
\item{\code{protected_resources}}{Protected resources URL.}
}
\description{
Hattrick OAuth URLs.
}
\examples{
# Request token URL
oauth_url()$request_token
# Authorize URL
oauth_url()$authorize
# Access token URL
oauth_url()$access_token
}
