% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow.R
\name{protected_resources}
\alias{protected_resources}
\title{Protected Resources}
\usage{
protected_resources(query, consumer, access_token, parse = FALSE)
}
\arguments{
\item{query}{A query string, e.g. \code{"file=teamdetails&version=3.4"}.}

\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{access_token}{An access token, created by \code{\link{access_token}}.}

\item{parse}{Parse the result using \code{\link[xml2]{read_xml}}?}
}
\value{
If \code{parse} is \code{FALSE}, a character vector, otherwise an
object of class \code{"xml_document"}.
}
\description{
Get protected resources (CHPP files in XML format).
}
\examples{
\dontrun{
# Consumer key and secret below are for illustrative purposes only
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu")
rtkn <- request_token(cns)
aurl <- authorization_url(cns, rtkn)
# Open in browser and enter your login name and password
# to obtain your verification code
browseURL(aurl)
# Verification code below is for illustrative purposes only
atkn <- access_token(cns, rtkn, "CzYLOvq9NniD5L8T")
pres <- protected_resources("file=teamdetails&version=3.4", cns, atkn)
}
}
