% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{request_token}
\alias{request_token}
\title{OAuth Request Token}
\usage{
request_token(consumer, callback = "oob")
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{callback}{Callback URL, \code{"oob"} by default.}
}
\value{
A request token, an object of class \code{"token"}.
}
\description{
Get a request token.
}
\examples{
\dontrun{
# consumer key and secret below are for illustrative purposes only
my_consumer <- consumer(
  "WdGMxbWypO3baqGSkrZGtX",
  "Bp7dG0GVFJNkAUPT6rHuptmlt1JRqJVIC46ElPHxejf"
)
my_request_token <- request_token(my_consumer)
}
}
