% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-constructors.R
\name{token}
\alias{token}
\title{OAuth Token}
\usage{
token(token, secret)
}
\arguments{
\item{token}{Token.}

\item{secret}{Token secret.}
}
\value{
An object of class \code{"token"}.
}
\description{
Create a token.
}
\examples{
# token and token secret below are for illustrative purposes only
my_token <- token("pCJS5wFvFfJWVhtY", "irhfsbr7C8c3OwtT")
}
