% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow.R
\name{access_token}
\alias{access_token}
\title{Access Token}
\usage{
access_token(consumer, request_token, verifier)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{request_token}{A request token, created by \code{\link{request_token}}.}

\item{verifier}{A verification code obtained after successful authorization.}
}
\value{
An access token, an object of class \code{"token"}.
}
\description{
Get an access token.
}
\examples{
\dontrun{
# Consumer key and secret below are for illustrative purposes only
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu")
rtkn <- request_token(cns)
aurl <- authorization_url(cns, rtkn)
# Open in browser and enter your login name and password
# to obtain your verification code
browseURL(aurl)
# Verification code below is for illustrative purposes only
atkn <- access_token(cns, rtkn, "CzYLOvq9NniD5L8T")
}
}
