% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow.R
\name{authorization_url}
\alias{authorization_url}
\title{Authorization URL}
\usage{
authorization_url(consumer, request_token)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}

\item{request_token}{A request token, created by \code{\link{request_token}}.}
}
\value{
An authorization URL.
}
\description{
Build an authorization URL.
}
\examples{
\dontrun{
# Consumer key and secret below are for illustrative purposes only
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu")
rtkn <- request_token(cns)
aurl <- authorization_url(cns, rtkn)
}
}
