% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-consumer.R
\name{consumer}
\alias{consumer}
\title{OAuth Consumer}
\usage{
consumer(key, secret, callback = "oob", challenge = FALSE, order = FALSE,
  youth = FALSE, training = FALSE, bid = FALSE)
}
\arguments{
\item{key}{Consumer key.}

\item{secret}{Consumer secret.}

\item{callback}{Callback URL (\code{"oob"} by default).}

\item{challenge}{Manage challenges?}

\item{order}{Set match orders?}

\item{youth}{Manage youth players?}

\item{training}{Set training?}

\item{bid}{Place bid?}
}
\value{
An object of class \code{"consumer"}.
}
\description{
Create a CHPP product (a consumer in OAuth terminology).
}
\examples{
# Consumer key and secret below are for illustrative purposes only
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu")
# Ask for permission to manage challenges
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu",
                challenge = TRUE)
}
