% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow.R
\name{request_token}
\alias{request_token}
\title{Request Token}
\usage{
request_token(consumer)
}
\arguments{
\item{consumer}{A consumer, created by \code{\link{consumer}}.}
}
\value{
A request token, an object of class \code{"token"}.
}
\description{
Get a request token.
}
\examples{
\dontrun{
# Consumer key and secret below are for illustrative purposes only
cns <- consumer("3qxJ4m36ONdmkQxVES9xW5",
                "nOhqxax1vDLElZPXgSzYOWIGWbDTQD1BpegtGPz4sCu")
rtkn <- request_token(cns)
}
}
