% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-token.R
\name{token}
\alias{token}
\title{OAuth Token}
\usage{
token(token, secret)
}
\arguments{
\item{token}{Token.}

\item{secret}{Token secret.}
}
\value{
An object of class \code{"token"}.
}
\description{
Create a token.
}
\examples{
# Token and secret below are for illustrative purposes only
tkn <- token("OpDVf2v0vuD31yW7", "ASyuUmShpiodN2WX")
}
